COLDWEB-5.1 Refactoring
